Refactor memory management; convert Anatomy and Health structs to ref structs, optimize GameContext and WarSnake classes, and introduce ColdMemoryLayout and HotMemoryLayout for improved memory layout and access

+++ Thanos/SourceGen/ThanosSerializerContext.cs
@@ -152,13 +152,13 @@ public readonly struct Coordinate(int x, int y)
 }
 
 [method: JsonConstructor]
-public readonly struct Snake(string id, string name, int health, Coordinate[] body, string latency, Coordinate head, int length, string shout)
+public readonly struct Snake(string id, string name, byte health, Coordinate[] body, string latency, Coordinate head, ushort length, string shout)
 {
     [JsonPropertyName("id")] public string Id { get; } = id;
 
     [JsonPropertyName("name")] public string Name { get; } = name;
 
-    [JsonPropertyName("health")] public int Health { get; } = health;
+    [JsonPropertyName("health")] public byte Health { get; } = health;
 
     [JsonPropertyName("body")] public Coordinate[] Body { get; } = body;
 
@@ -166,7 +166,7 @@ public readonly struct Snake(string id, string name, int health, Coordinate[] bo
 
     [JsonPropertyName("head")] public Coordinate Head { get; } = head;
 
-    [JsonPropertyName("length")] public int Length { get; } = length;
+    [JsonPropertyName("length")] public ushort Length { get; } = length;
 
     [JsonPropertyName("shout")] public string Shout { get; } = shout;
